requirements.txt
python-telegram-bot
telegram
numpy
matplotlib.pyplot
networkx